requirements.txt
streamlit
torch
torchvision
transformers
faiss-cpu
opencv-python
Pillow
matplotlib
numpy
